requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aget==0.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alabaster==0.7.16
AliPCS-Py==0.8.1
altair==5.5.0
amdsmi @ file:///opt/rocm-6.3.0/share/amd_smi
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
apex @ file:///var/lib/jenkins/apex
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
autocommand==2.2.2
av==14.1.0
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
azure-storage-file-datalake==12.18.1
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bitsandbytes @ file:///InstellaVL/bitsandbytes
bleach==6.2.0
blinker==1.9.0
blobfile==3.0.0
boto3==1.35.63
botocore==1.35.97
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
circuitbreaker==2.0.0
click==8.1.7
cmake==3.31.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coremltools==5.0b5
coverage==7.6.12
cramjam==2.9.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
databricks-sdk==0.29.0
DataProperty==1.1.0
datasets==2.21.0
debugpy==1.8.12
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.4
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==5.1.4
docformatter==1.7.5
docker-pycreds==0.4.0
docutils==0.17.1
ecdsa==0.19.0
einops==0.8.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.2.0
expecttest==0.1.6
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flash_attn==3.0.0.post1
flatbuffers==2.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
furo==2022.9.29
future==1.0.0
geojson==2.5.0
ghstack==0.8.0
gitdb==4.0.12
GitPython==3.1.42
google-api-core==2.24.1
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.2
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
gql==3.5.1
graphql-core==3.2.3
grouped_gemm @ file:///workspace/grouped_gemm
grpcio==1.68.1
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanize==4.12.1
hypothesis==5.35.1
identify==2.6.7
idna==3.10
image==1.5.33
imageio==2.36.1
imagesize==1.4.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junitparser==2.1.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.8
kiwisolver==1.4.8
lark==0.12.0
lazy_loader==0.4
librosa==0.10.2.post1
lintrunner==0.12.5
lion-pytorch==0.1.4
llvmlite==0.38.1
loguru==0.7.3
lxml==5.0.0
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.3.0
mosaicml-cli==0.7.2
-e git+https://github.com/mosaicml/streaming.git@6af3216e5608899e97f2f75be47c21d4106392fb#egg=mosaicml_streaming
moto==5.0.28
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
myst-parser==0.16.1
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.1
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1.3
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.55.2
numpy==1.26.4
oauthlib==3.2.2
oci==2.145.0
omegaconf==2.3.0
openai==1.63.2
openpyxl==3.1.5
opt-einsum==3.3.0
optionloop==1.0.7
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pathvalidate==3.2.3
peewee==3.17.9
peft==0.12.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
ply==3.11
pooch==1.8.2
portalocker==2.10.1
pre-commit==3.8.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==3.20.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
PyGithub==2.3.0
Pygments==2.15.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pypandoc==1.13
pyparsing==3.2.1
pyspark==3.5.4
pytablewriter==1.2.1
pytest==8.2.2
pytest-cov==6.0.0
pytest-cpp==2.3.0
pytest-flakefinder==1.1.0
pytest-rerunfailures==14.0
pytest-split==0.10.0
pytest-xdist==3.3.1
pytest_codeblocks==0.17.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-snappy==0.7.3
pytz==2024.2
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
qrcode==8.0
questionary==2.1.0
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.6
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rockset==1.0.3
rpds-py==0.23.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
sacrebleu==2.5.1
safetensors==0.5.2
scikit-image==0.20.0
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
Sphinx==4.4.0
sphinx-argparse==0.4.0
sphinx-basic-ng==1.0.0b2
sphinx-copybutton==0.5.2
sphinx-tabs==3.4.5
sphinxcontrib-applehelp==1.0.0
sphinxcontrib-devhelp==1.0.0
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-katex==0.9.6
sphinxcontrib-qthelp==1.0.0
sphinxcontrib-serializinghtml==1.1.5
sqlitedict==2.1.0
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.41.3
streamlit==1.42.2
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tb-nightly==2.13.0a20230426
tcolorpy==0.1.7
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.12
tinycss2==1.4.0
tlparse==0.3.7
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0a0+git6374332
torchao==0.9.0.dev20250214+rocm6.3
# Editable install with no version control (torchtune==0.0.0)
-e /workspace/torchtune
torchvision==0.22.0a0+867521e
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformer_engine @ file:///var/lib/jenkins/TransformerEngine
transformers==4.44.0
triton==3.1.0
typeguard==4.3.0
typepy==1.3.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
unittest-xml-reporting==3.2.0
untokenize==0.1.1
uri-template==1.3.0
urllib3==1.26.20
uv==0.5.24
uvicorn==0.34.0
validators==0.34.0
virtualenv==20.29.2
wandb==0.19.7
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xdoctest==1.1.0
xmltodict==0.14.2
xxhash==3.5.0
yamllint==1.35.1
yarl==1.18.3
z3-solver==4.12.2.0
zipp==3.19.2
zstd==1.5.6.3
